업로드된 각 파일들에 대해서 메타정보(GenFile) 생성

+++ fileUpload/src/main/java/com/ll/exam/fileUpload/app/base/entity/GenFile.java
@@ -14,7 +14,7 @@ import javax.persistence.Entity;
 @ToString(callSuper = true)
 public class GenFile extends BaseEntity {
     private String relTypeCode;
-    private int relId;
+    private Long relId;
     private String typeCode;
     private String type2Code;
     private String fileExtTypeCode;
